% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_stats.R
\name{celltype_features}
\alias{celltype_features}
\title{Retrieve stats (sum, n_cells, avg) on transcripts expressed in a given cell type
Transcripts with zero expression are filtered out.}
\usage{
celltype_features(obj, count_matrix, cell_type, features = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{count_matrix}{assay count matrix}

\item{cell_type}{Identity}

\item{features}{(optional) subset rows in matrix}
}
\value{
data.frame with features expressed in a cell type
(feature/gene_id/transcript_id) + cell_type, sum, n_cells, avg
}
\description{
Retrieve stats (sum, n_cells, avg) on transcripts expressed in a given cell type
Transcripts with zero expression are filtered out.
}
