% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_stats.R
\name{celltype_features_bulk}
\alias{celltype_features_bulk}
\title{Compute number of genes and transcripts per cell type using pseudo-bulk}
\usage{
celltype_features_bulk(obj, assay, bulk_bin_size = 10)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to pull data from}

\item{bulk_bin_size}{Size of the pseudobulk groups, defaults to 10.}
}
\value{
Dataframe with number of genes and number of transcripts found in each bin
}
\description{
Compute number of genes and transcripts per cell type using pseudo-bulk
}
