% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{compute_switches}
\alias{compute_switches}
\title{Computes isoform switches for a specific gene in a marker list}
\usage{
compute_switches(
  marker_list,
  cutoff = 0.05,
  gene = NULL,
  cluster = NULL,
  order = NULL
)
}
\arguments{
\item{marker_list}{Marker list from ISO_SWITCH_ALL}

\item{cutoff}{tbc}

\item{gene}{(optional)}

\item{cluster}{(optional)}

\item{order}{(optional)}
}
\value{

}
\description{
Computes isoform switches for a specific gene in a marker list
}
