% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{._gene_switches}
\alias{._gene_switches}
\title{Internal worker function for compute_switches}
\usage{
._gene_switches(gene, marker_list, cutoff = 0.05)
}
\arguments{
\item{gene}{Gene filter}

\item{marker_list}{input marker list as returned by ISO_SWITCH_ALL}

\item{cutoff}{unused}
}
\value{
data.frame with isoform switches
}
\description{
Internal worker function for compute_switches
}
