% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{._isoswitch_report.dotpot}
\alias{._isoswitch_report.dotpot}
\title{Title}
\usage{
._isoswitch_report.dotpot(
  obj,
  obj_assay,
  meta,
  celltype_order = NULL,
  switch = NULL
)
}
\arguments{
\item{obj}{tbc}

\item{obj_assay}{tbc}

\item{meta}{tbc}

\item{celltype_order}{tbc}

\item{switch}{tbc}
}
\value{
tbc
}
\description{
Title
}
