% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{._isoswitch_report.junctions}
\alias{._isoswitch_report.junctions}
\title{Title}
\usage{
._isoswitch_report.junctions(obj, assay, gtf_df, gene, meta)
}
\arguments{
\item{obj}{tbc}

\item{assay}{tbc}

\item{gtf_df}{tbc}

\item{gene}{tbc}

\item{meta}{tbc}
}
\value{
tbc
}
\description{
Title
}
