% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{._isoswitch_report.locus}
\alias{._isoswitch_report.locus}
\title{Title}
\usage{
._isoswitch_report.locus(gtf_df, gene, meta, legend = TRUE)
}
\arguments{
\item{gtf_df}{tbc}

\item{gene}{tbc}

\item{meta}{tbc}

\item{legend}{tbc}
}
\value{
tbc
}
\description{
Title
}
