% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{._isoswitch_report.umi_counts}
\alias{._isoswitch_report.umi_counts}
\title{Stacked barplot representing average raw transcript counts per cell for each cell type}
\usage{
._isoswitch_report.umi_counts(obj, gene, meta, legend = FALSE)
}
\arguments{
\item{obj}{Seurat object}

\item{gene}{Gene}

\item{meta}{Metadata computed by isoswitch_report function}

\item{legend}{Legend boolead, defaults to False for iso_dotplo2 reports}
}
\value{
ggplot object
}
\description{
Stacked barplot representing average raw transcript counts per cell for each cell type
}
