% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{._order_switches}
\alias{._order_switches}
\title{Internal helper function to select the criteria used to sort the list of switches
centralizing it here, to avoid duplicating code}
\usage{
._order_switches(switch_list, order = NULL)
}
\arguments{
\item{switch_list}{switch list}

\item{order}{One of "minp_not_major" (default), "max_p", "diff.pct"}
}
\value{
sorted dataframe according to input criteria
}
\description{
Internal helper function to select the criteria used to sort the list of switches
centralizing it here, to avoid duplicating code
}
