% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_stats.R
\name{format_switch_table}
\alias{format_switch_table}
\title{Clean up / format switch data frame for reports}
\usage{
format_switch_table(switch_df)
}
\arguments{
\item{switch_df}{Output from call to compute_switches}
}
\value{
same df with selected columns and number formatting
}
\description{
Clean up / format switch data frame for reports
}
