% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{gene_switch_table}
\alias{gene_switch_table}
\title{Returns a html switch table for reports}
\usage{
gene_switch_table(marker_list)
}
\arguments{
\item{marker_list}{Marker list returned by ISO_SWITCH_ALL}
}
\value{
A reactable printable object
}
\description{
Can only be used on html output documents.
}
