% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_preprocess.R
\name{iso_preprocess}
\alias{iso_preprocess}
\title{Preprocess isoform-level count matrix for isoform switch.}
\usage{
iso_preprocess(obj, assay, new_assay, filter_threshold)
}
\arguments{
\item{obj}{Input Seurat object}

\item{assay}{Assay containing unfiltered isoform-level counts}

\item{new_assay}{Name of the new assay to store pre-processed counts}

\item{filter_threshold}{Filtering threshold expressed as percentage of total gene expression}
}
\value{
Input Seurat object with additional assay contaning pre-processed counts
}
\description{
Removes transcripts with low relative expression compared to the total gene
expression. Filtering threshold is given as parameter. Single-isoform genes
(after the filtering step) are also removed. The resulting matrix is stored
as a new assay in the input Seurat object
}
