% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{ISO_SWITCH_ALL}
\alias{ISO_SWITCH_ALL}
\title{Find transcript markers of isoform switches}
\usage{
ISO_SWITCH_ALL(seurat, clusters, assay, ...)
}
\arguments{
\item{seurat}{Seurat object}

\item{clusters}{Clusters to compare}

\item{assay}{Assay to pull data from}

\item{...}{extra parameters passed on to Seurat FindMarkers call}
}
\value{
Marker list data frame
}
\description{
Find transcript markers of isoform switches
}
