% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{isoswitch_report}
\alias{isoswitch_report}
\title{Gene multi-panel report, including gene locus and junction quantification,
average UMI counts per isoform (p1), dotplot on scale data (p2) and violin
plots of the switch. For genes with several switches the first one is
selected by default, unless overridden with parameter i.}
\usage{
isoswitch_report(obj, obj_assay, marker_list, gene, i = 1)
}
\arguments{
\item{obj}{Seurat object}

\item{obj_assay}{Assay to pull data from}

\item{marker_list}{Output from ISO_SWITCH_ALL call}

\item{gene}{Gene to report on}

\item{i}{index of switch to represent}
}
\value{
ggplot object
}
\description{
Gene multi-panel report, including gene locus and junction quantification,
average UMI counts per isoform (p1), dotplot on scale data (p2) and violin
plots of the switch. For genes with several switches the first one is
selected by default, unless overridden with parameter i.
}
