% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_report.R
\name{isoswitch_report_short}
\alias{isoswitch_report_short}
\title{Stripped-down version of isoswitch_report with only barplot}
\usage{
isoswitch_report_short(obj, obj_assay, marker_list, gene, i = 1)
}
\arguments{
\item{obj}{Seurat object}

\item{obj_assay}{Assay to pull data from}

\item{marker_list}{Output from ISO_SWITCH_ALL call}

\item{gene}{Gene to report on}

\item{i}{(unused)}
}
\value{
tbc
}
\description{
Stripped-down version of isoswitch_report with only barplot
}
