% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_visual.R
\name{plot_assay_stats}
\alias{plot_assay_stats}
\title{Plot summary stats on gene and isoform expression}
\usage{
plot_assay_stats(obj, assay, bin_size = 10)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to pull data from}

\item{bin_size}{Bin size for pseudobulk aggregation of number of genes per cell}
}
\value{
ggplot object, patchwork of 4 plots
}
\description{
Produces a 4 panel plot including a) Number of total genes and transcripts
and percentage of genes expressing one or more isoforms. (b) Distribution
of the number of isoforms per gene. (c) Relationship between the number of
major isoforms (y-axis) and the total number of isoforms per gene (x-axis).
(d) Pseudo-bulk estimation of number of genes expressed per cell type
(each data point represents the aggregation of 10 cells)
}
