% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_visual.R
\name{plot_marker_matrix}
\alias{plot_marker_matrix}
\title{Plot matrix of genes found per contrast}
\usage{
plot_marker_matrix(obj, marker_list)
}
\arguments{
\item{obj}{Seurat object}

\item{marker_list}{Marker list returned by ISO_SWITCH_ALL}
}
\value{
ggplot object
}
\description{
Plot matrix of genes found per contrast
}
