% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_visual.R
\name{plot_marker_score}
\alias{plot_marker_score}
\title{Volcano-like plot showing log_f2c vs p-value for genes found in marker_list
returned by ISO_SWITCH_ALL}
\usage{
plot_marker_score(obj, markers, facet = FALSE, ncol = 4, overlaps = 10)
}
\arguments{
\item{obj}{Seurat object}

\item{markers}{List of markers returned by ISO_SWITCH_ALL}

\item{facet}{if TRUE, facet by individual cell type, defaults to FALSE}

\item{ncol}{specifies number of columns if facet=TRUE, defaults to 4}

\item{overlaps}{max number of overlapping labels, defaults to 10}
}
\value{
ggplot object
}
\description{
Volcano-like plot showing log_f2c vs p-value for genes found in marker_list
returned by ISO_SWITCH_ALL
}
