% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_switches.R
\name{scale_my_data}
\alias{scale_my_data}
\title{Scales raw transcript counts relative to total gene expression}
\usage{
scale_my_data(seurat, assay, name = "scale.data")
}
\arguments{
\item{seurat}{Seurat object}

\item{assay}{Assay to pull data from}

\item{name}{Name for new assay with scaled data}
}
\value{
Seurat object with scaled data in new slot
}
\description{
Scales raw transcript counts relative to total gene expression
}
