% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_visual.R
\name{plot_assay_stats}
\alias{plot_assay_stats}
\title{Title}
\usage{
plot_assay_stats(obj, assay)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to pull data from}
}
\value{
ggplot object, patchwork of 4 plots
}
\description{
Title
}
